backend/requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.1.0
Flask-Cors==5.0.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
packaging==24.2
prometheus_client==0.21.1
prometheus_flask_exporter==0.23.1
python-dotenv==1.0.1
setuptools==75.1.0
Werkzeug==3.1.3
wheel==0.44.0
